Re-add proper test class
The base is now abstract.
